+++ src/Function/List/LstMapFunction.php
@@ -34,8 +34,8 @@ final class LstMapFunction extends ListMapFunction {
 	/**
 	 * @inheritDoc
 	 */
-	public function allowsNamedParams(): bool {
-		return false;
+	public function getParserFlags(): int {
+		return 0;
 	}
 
 	/**