+++ src/Function/List/LstMapTempFunction.php
@@ -19,8 +19,8 @@ final class LstMapTempFunction extends ListMapFunction {
 	/**
 	 * @inheritDoc
 	 */
-	public function allowsNamedParams(): bool {
-		return false;
+	public function getParserFlags(): int {
+		return 0;
 	}
 
 	/**