+++ src/Function/List/LstSrtFunction.php
@@ -19,8 +19,8 @@ final class LstSrtFunction extends ListSortFunction {
 	/**
 	 * @inheritDoc
 	 */
-	public function allowsNamedParams(): bool {
-		return false;
+	public function getParserFlags(): int {
+		return 0;
 	}
 
 	/**