+++ src/Function/List/LstUniqFunction.php
@@ -19,8 +19,8 @@ final class LstUniqFunction extends ListUniqueFunction {
 	/**
 	 * @inheritDoc
 	 */
-	public function allowsNamedParams(): bool {
-		return false;
+	public function getParserFlags(): int {
+		return 0;
 	}
 
 	/**