+++ src/ParameterParser.php
@@ -6,6 +6,7 @@ namespace MediaWiki\Extension\ParserPower;
 
 use MediaWiki\Parser\Parser;
 use MediaWiki\Parser\PPFrame;
+use MediaWiki\Parser\PPNode;
 
 /**
  * Parameter parser for parser functions.
@@ -22,6 +23,11 @@ final class ParameterParser {
 	 * Flag for whether named arguments are allowed, and should be split from numbered arguments.
 	 */
 	public const ALLOWS_NAMED = 1;
+	/**
+	 * Flag for tracking parameter values that contain an = at the root level, that may be confused as a parameter key.
+	 * May be used for migration purpose before enabling ALLOWS_NAMED.
+	 */
+	public const TRACKS_NAMED_VALUES = 2;
 
 	/**
 	 * @param array $paramOptions Parsing and post-processing options for all parameters.
@@ -33,6 +39,26 @@ final class ParameterParser {
 		private ?array $defaultOptions = null,
 		private int $flags = 0
 	) {
+	}
+
+	/**
+	 * Split a parameter key from its value.
+	 *
+	 * @param string|PPNode $param Parameter to split.
+	 * @return array Pair with the parameter key (if named) and value.
+	 */
+	private function splitKeyValue( string|PPNode $param ): array {
+		if ( is_string( $param ) ) {
+			$pair = explode( '=', $param, 2 );
+			$key = isset( $pair[1] ) ? array_shift( $pair ) : null;
+			$value = $pair[0];
+		} else {
+			$bits = $param->splitArg();
+			$key = $bits['index'] === '' ? $bits['name'] : null;
+			$value = $bits['value'];
+		}
+
+		return [ $key, $value ];
 	}
 
 	/**
@@ -50,15 +76,7 @@ final class ParameterParser {
 		foreach ( $rawParams as $rawParam ) {
 			// Split key from value
 			if ( $this->flags & self::ALLOWS_NAMED ) {
-				if ( is_string( $rawParam ) ) {
-					$pair = explode( '=', $rawParam, 2 );
-					$key = isset( $pair[1] ) ? array_shift( $pair ) : null;
-					$value = $pair[0];
-				} else {
-					$bits = $rawParam->splitArg();
-					$key = $bits['index'] === '' ? $bits['name'] : null;
-					$value = $bits['value'];
-				}
+				[ $key, $value ] = $this->splitKeyValue( $rawParam );
 			} else {
 				$key = null;
 				$value = $rawParam;
@@ -78,6 +96,18 @@ final class ParameterParser {
 			} else {
 				$parser->addTrackingCategory( 'parserpower-invalid-args-category' );
 				continue;
+			}
+
+			if ( $key === null && $this->flags & self::TRACKS_NAMED_VALUES ) {
+				try {
+					[ $subKey, ] = $this->splitKeyValue( $value );
+				} catch ( InvalidArgumentException ) {
+					// Not a parameter, probably not an issue.
+					$subKey = null;
+				}
+				if ( $subKey !== null ) {
+					$parser->addTrackingCategory( 'parserpower-invalid-args-category' );
+				}
 			}
 
 			// Resolve aliases